streamlit/requirements.txt
# Streamlit app dependencies
streamlit
pillow
numpy
pandas